Alias tool spring-mvc support.
This is a proof of concept support of spring mvc inside the alias tool.
This allows you to develop new functionality with spring mvc while continuing to support the old servlet. Basically the old servlet is the default handler in spring MVC so it first looks for supported annotated controllers and if it doesn’t find any it falls back to the standard tool.

To have the hello world work just append /helloWorld to the URL inside the aliases too.